Small refactors to tracklist
